// mobile/src/types/propsType.ts
export type ColorType =
  | 'primary'
  | 'primaryLight'
  | 'primaryDark'
  | 'secondary'
  | 'success'
  | 'warning'
  | 'danger'
  | 'grayDark'
  | 'grayLight'
  | 'white';
export type SizeText = 'large' | 'medium' | 'small';
export type WeightText = 'bold' | 'normal' | 'light';
export type OpacityType = 'primary' | 'secondary' | 'tertiary';